Finish blog  module

+++ services/committee/committee.service.js
@@ -1,6 +1,7 @@
 const CoreService = require("../core.service.js");
 const Committee = require("../../models/Committee.model");
 const User = require("../../models/User.model");
+const BestMember = require("../../models/Best-member.model");
 
 class CommitteeService extends CoreService {
   constructor() {
@@ -21,18 +22,27 @@ class CommitteeService extends CoreService {
               "There's users on this committee, please remove users first then retry !",
           });
         } else {
-          this.db
-            .findByIdAndRemove(id)
-            .then(() => {
-              res.json({
-                msg: `${this.name} has been deleted successfully!`,
+          BestMember.find({ committee: committee.name }).then((users) => {
+            if (users && users.length > 0) {
+              res.status(400).json({
+                msg:
+                  "There's members on this committee, please remove members first then retry !",
               });
-            })
-            .catch(() =>
-              res.status(500).json({
-                msg: "An error occurred, please try again later!",
-              })
-            );
+            } else {
+              this.db
+                .findByIdAndRemove(id)
+                .then(() => {
+                  res.json({
+                    msg: `${this.name} has been deleted successfully!`,
+                  });
+                })
+                .catch(() =>
+                  res.status(500).json({
+                    msg: "An error occurred, please try again later!",
+                  })
+                );
+            }
+          });
         }
       });
     });
